refactor: add file input hook
